requirements.txt
streamlit
PyPDF2
faiss-cpu
mistralai
numpy
mysql-connector-python
bcrypt